+++ packages/abap-deploy-config-writer/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/abap-deploy-config-writer
 
+## 0.0.69
+
+### Patch Changes
+
+-   @sap-ux/system-access@0.5.23
+-   @sap-ux/ui5-config@0.26.0
+
+## 0.0.68
+
+### Patch Changes
+
+-   1bb4d48: adds new module @sap-ux/abap-deploy-config-sub-generator
+
+## 0.0.67
+
+### Patch Changes
+
+-   @sap-ux/system-access@0.5.22
+
+## 0.0.66
+
+### Patch Changes
+
+-   Updated dependencies [e93797a]
+    -   @sap-ux/project-access@1.28.9
+    -   @sap-ux/system-access@0.5.21
+
 ## 0.0.65
 
 ### Patch Changes