+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/telemetry
 
+## 0.5.48
+
+### Patch Changes
+
+-   Updated dependencies [3ebd767]
+    -   @sap-ux/store@1.0.0
+    -   @sap-ux/ui5-config@0.26.0
+
+## 0.5.47
+
+### Patch Changes
+
+-   Updated dependencies [e93797a]
+    -   @sap-ux/project-access@1.28.9
+
 ## 0.5.46
 
 ### Patch Changes