+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.8.11
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.5.12
+
+## 0.8.10
+
+### Patch Changes
+
+-   Updated dependencies [0c64478]
+    -   @sap-ux/inquirer-common@0.5.11
+
+## 0.8.9
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.5.10
+
+## 0.8.8
+
+### Patch Changes
+
+-   Updated dependencies [e93797a]
+    -   @sap-ux/project-access@1.28.9
+    -   @sap-ux/inquirer-common@0.5.9
+
 ## 0.8.7
 
 ### Patch Changes