+++ packages/ui5-library-reference-inquirer/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/ui5-library-reference-inquirer
 
+## 0.3.57
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.5.12
+
+## 0.3.56
+
+### Patch Changes
+
+-   Updated dependencies [0c64478]
+    -   @sap-ux/inquirer-common@0.5.11
+
+## 0.3.55
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.5.10
+
+## 0.3.54
+
+### Patch Changes
+
+-   Updated dependencies [e93797a]
+    -   @sap-ux/project-access@1.28.9
+    -   @sap-ux/inquirer-common@0.5.9
+
 ## 0.3.53
 
 ### Patch Changes